reorganize!
we may eventuall want to make more methods on SuperSchema
